requirement.txt
streamlit 
agno
openai 
duckduckgo-search 
yfinance
python-dotenv
packaging
fastapi
uvicorn
google-genai